feat(gui): 添加桌面环境和窗口管理器框架
- 添加桌面环境模块 (desktop/mod.rs)
- 修复窗口管理器编译问题 (BTreeMap 替代 BTreeSet)
- 修复 PS/2 键盘驱动类型问题 (u16 替代 u8)
- 暂时禁用图形系统初始化 (需要 VirtIO-GPU 支持)
- 修复输入事件队列类型问题

注意：图形系统暂时禁用，等待 VirtIO-GPU 驱动实现

+++ kernel/src/drivers/gpu/framebuffer.rs
@@ -222,7 +222,7 @@ impl FrameBuffer {
     pub fn draw_bitmap(&self, x: u32, y: u32, width: u32, height: u32, data: &[u8], color: u32) {
         for py in 0..height {
             for px in 0..width {
-                let byte_index = (py * width + px) / 8;
+                let byte_index = ((py * width + px) / 8) as usize;
                 let bit_index = 7 - ((py * width + px) % 8);
 
                 if byte_index < data.len() {
